microservices/academic-service/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
drf-spectacular==0.26.5
django-filter==23.3
psycopg2-binary==2.9.7
gunicorn==21.2.0
python-decouple==3.8
